9 готовая с маршрутизатором

+++ src/features/CityCardView/CityCardView.tsx
@@ -1,8 +1,8 @@
 // src/features/CityCardView/CityCardView.tsx
+import { Link } from "react-router-dom";
 import { useCities } from "../../contexts/CityContext";
-import { useNavigation } from "../../contexts/NavigationContext";
 import { ICity } from "../../entities/City/model";
-import  CityCard  from "../../entities/City/CityCard";
+import { CityCard } from "../../entities/City/CityCard";
 import Button from "../../shared/ui/Button";
 import FlexContainer from "../../shared/ui/FlexContainer";
 import Panel from "../../shared/ui/Panel";
@@ -12,22 +12,25 @@ interface ICityCardViewProps {
 }
 
 export function CityCardView({ city }: ICityCardViewProps) {
-    const { deleteCity } = useCities();
-    const { goToEdit, goToDetails } = useNavigation();
+    const { deleteCity } = useCities(); // Используем хук useCities
 
     const handleDelete = () => {
-        deleteCity(city.id);
+        if (window.confirm(`Вы уверены, что хотите удалить город ${city.name}?`)) {
+            deleteCity(city.id);
+        }
     };
 
     return (
         <Panel padding={5}>
             <FlexContainer>
-                <div onClick={() => goToDetails(city.id)} style={{ cursor: "pointer", flex: 1 }}>
+                <Link to={`/city/${city.id}`} style={{ textDecoration: "none", flex: 1 }}>
                     <CityCard city={city} />
-                </div>
+                </Link>
                 <FlexContainer direction="row">
                     <Button clickAction={handleDelete}>Удалить</Button>
-                    <Button clickAction={() => goToEdit(city.id)}>Изменить</Button>
+                    <Link to={`/edit/${city.id}`}>
+                        <Button>Изменить</Button>
+                    </Link>
                 </FlexContainer>
             </FlexContainer>
         </Panel>
